price_scraper/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.3.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0 